+++ src/common/src/hash/consistent_hash/mod.rs
@@ -1,0 +1,2 @@
+pub mod mapping;
+pub mod vnode;
