+++ src/common/src/hash/mod.rs
@@ -12,11 +12,11 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-mod key;
-pub use key::*;
+mod consistent_hash; // TODO: move this to a separate module
 mod dispatcher;
+mod key;
+
+pub use consistent_hash::mapping::*;
+pub use consistent_hash::vnode::*;
 pub use dispatcher::HashKeyDispatcher;
-mod vnode;
-pub use vnode::*;
-mod mapping;
-pub use mapping::*;
+pub use key::*;
