+++ src/meta/src/stream/mod.rs
@@ -12,7 +12,6 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-mod mapping;
 mod scale;
 mod scheduler;
 mod source_manager;
@@ -22,7 +21,6 @@ mod stream_manager;
 mod test_fragmenter;
 mod test_scale;
 
-pub use mapping::*;
 pub use scale::*;
 pub use scheduler::*;
 pub use source_manager::*;
